Refactor contracts to lib - it uses supabase and firebase
