+++ web/lib/supabase/contracts.ts
@@ -1,35 +1,11 @@
-import { Bet } from 'common/bet'
-import { getInitialProbability } from 'common/calculate'
+import { CPMMContract, visibility, Contract } from 'common/contract'
 import {
-  BinaryContract,
-  CPMMContract,
-  PseudoNumericContract,
-  visibility,
-} from 'common/contract'
-import { getTotalContractMetrics } from 'common/supabase/contract-metrics'
-import {
-  run,
   millisToTs,
+  run,
   selectJson,
   SupabaseClient,
 } from 'common/supabase/utils'
 import { filterDefined } from 'common/util/array'
-import { removeUndefinedProps } from 'common/util/object'
-import { compressPoints, pointsToBase64 } from 'common/util/og'
-import { HistoryPoint } from 'web/components/charts/generic-charts'
-import { getRelatedContracts } from 'web/hooks/use-related-contracts'
-import { getUser } from 'web/lib/supabase/user'
-import {
-  ContractParams,
-  CONTRACT_BET_FILTER,
-} from 'web/pages/[username]/[contractSlug]'
-import {
-  getBinaryContractUserContractMetrics,
-  getTopContractMetrics,
-} from '../firebase/contract-metrics'
-import { Contract } from '../firebase/contracts'
-import { getBets, getTotalBetCount } from './bets'
-import { getAllComments } from './comments'
 import { adminDb, db } from './db'
 
 export async function getContractIds(contractIds: string[]) {
@@ -139,88 +115,6 @@ export async function getContractVisibilityFromSlug(contractSlug: string) {
       .visibility
   }
   return undefined
-}
-
-export async function getContractParams(contract: Contract) {
-  const contractId = contract.id
-  const totalBets = await getTotalBetCount(contractId)
-  const shouldUseBetPoints =
-    contract.outcomeType === 'BINARY' ||
-    contract.outcomeType === 'PSEUDO_NUMERIC'
-
-  // in original code, prioritize newer bets via descending order
-  const bets = await getBets({
-    contractId,
-    ...CONTRACT_BET_FILTER,
-    limit: shouldUseBetPoints ? 50000 : 4000,
-    order: 'desc',
-  })
-
-  const betPoints = shouldUseBetPoints
-    ? bets.map(
-        (bet) =>
-          removeUndefinedProps({
-            x: bet.createdTime,
-            y: bet.probAfter,
-            obj:
-              totalBets < 1000
-                ? { userAvatarUrl: bet.userAvatarUrl }
-                : undefined,
-          }) as HistoryPoint<Partial<Bet>>
-      )
-    : []
-
-  const comments = await getAllComments(contractId, 100)
-
-  const userPositionsByOutcome =
-    contract.outcomeType === 'BINARY'
-      ? await getBinaryContractUserContractMetrics(contractId, 100)
-      : {}
-
-  const topContractMetrics = contract.resolution
-    ? await getTopContractMetrics(contract.id, 10)
-    : []
-
-  const totalPositions =
-    contract.outcomeType === 'BINARY'
-      ? await getTotalContractMetrics(contractId, db)
-      : 0
-
-  if (shouldUseBetPoints) {
-    const firstPoint = {
-      x: contract.createdTime,
-      y: getInitialProbability(
-        contract as BinaryContract | PseudoNumericContract
-      ),
-    }
-    betPoints.push(firstPoint)
-    betPoints.reverse()
-  }
-
-  const pointsString =
-    contract.visibility != 'private'
-      ? pointsToBase64(compressPoints(betPoints))
-      : undefined
-
-  const creator = await getUser(contract.creatorId)
-
-  const relatedContracts = await getRelatedContracts(contract, 9)
-
-  return removeUndefinedProps({
-    contract,
-    historyData: {
-      bets: shouldUseBetPoints ? bets.slice(0, 100) : bets,
-      points: betPoints,
-    },
-    pointsString,
-    comments,
-    userPositionsByOutcome,
-    totalPositions,
-    totalBets,
-    topContractMetrics,
-    creatorTwitter: creator?.twitterHandle,
-    relatedContracts,
-  }) as ContractParams
 }
 
 export async function searchContract(query: string) {
